Better URL for thumbnails on Recent module
I've added the recent module to the home page of my
site. The picture thumbnails get an URL like this:

http://localhost/photolib/picture.php?/35851/category

When followed this lead to a blank page.

With this change the generated URL looks like this:

http://localhost/photolib/picture.php?/35851/recent_pics

which takes the user to the recent photos page - a better
user experience.

tjk :)